イベントビュアーの呼び出し側変更
GameScene→EventController
