+++ project/imgui.ini
@@ -29,7 +29,7 @@ Size=418,211
 Collapsed=0
 
 [Window][PlaceActor]
-Pos=60,60
+Pos=788,177
 Size=78,48
 Collapsed=0
