+++ project/source/Scene/GameScene.cpp
@@ -19,8 +19,6 @@
 #include "GameState/GameInit.h"
 #include "GameState/GameIdle.h"
 
-//※イベントコントローラーが出来たらそっち移動
-#include "../Viewer/GameScene/EventViewer/EventViewer.h"
 #include "../FieldObject/Actor/CrossJunctionActor.h"
 
 /**************************************
@@ -41,9 +39,6 @@ void GameScene::Init()
 	text = new TextViewer("M+ 2c heavy", 50);
 	gameViewer = new GameViewer();
 	eventController = new EventController(Field::Model::City);
-
-	//※イベントコントローラーが出来たらそっち移動
-	eventViewer = new EventViewer();
 
 	//ステートマシン作成
 	fsm.resize(State::Max, NULL);
@@ -73,9 +68,6 @@ void GameScene::Uninit()
 	SAFE_DELETE(text);
 	SAFE_DELETE(gameViewer);
 
-	//※イベントコントローラーが出来たらそっち移動
-	SAFE_DELETE(eventViewer);
-
 	//ステートマシン削除
 	Utility::DeleteContainer(fsm);
 
@@ -103,9 +95,6 @@ void GameScene::Update()
 
 	//ビュアー更新
 	gameViewer->Update();
-
-	//※イベントコントローラーが出来たらそっち移動
-	eventViewer->Update();
 
 	testActor->Update();
 }
@@ -136,9 +125,6 @@ void GameScene::Draw()
 
 	//ビュアー描画
 	gameViewer->Draw();
-
-	//※イベントコントローラーが出来たらそっち移動
-	eventViewer->Draw();
 
 	// イベント描画
 	eventController->Draw();
