+++ project/source/Scene/GameScene.h
@@ -30,9 +30,6 @@ class TextViewer;
 class GameViewer;
 class EventController;
 
-//※イベントコントローラーが出来たらそっち移動
-class EventViewer;
-
 /**************************************
 クラス定義
 ***************************************/
@@ -62,10 +59,6 @@ private:
 	Field::FieldController *field;				//フィールドコントローラ
 	TextViewer *text;							//テスト用テキストビューワ
 	GameViewer *gameViewer;						//ゲームビュアー
-
-	//※イベントコントローラーが出来たらそっち移動
-	EventViewer *eventViewer;
-
 
 	EventController *eventController;			// イベントコントローラー
 	float remainTime;							//ステージの残り時間
